Changed how assert works to break up the useless asm blocks

+++ Engine/Platform/System/CYBSyscalls.hpp
@@ -27,7 +27,7 @@ namespace CYB {
 				static unsigned long long DoCall(const CallNumber ACallNumber, const Union64 AArg1, const Union64 AArg2, const Union64 AArg3, const Union64 AArg4);
 				static unsigned long long DoCall(const CallNumber ACallNumber, const Union64 AArg1, const Union64 AArg2, const Union64 AArg3, const Union64 AArg4, const Union64 AArg5);
 				static unsigned long long DoCall(const CallNumber ACallNumber, const Union64 AArg1, const Union64 AArg2, const Union64 AArg3, const Union64 AArg4, const Union64 AArg5, const Union64 AArg6);
-				static void VerifyArgumentCount(const CallNumber ACallNumber, const size_t ANumArgs);
+				static void VerifyArgumentCount(const CallNumber ACallNumber, const unsigned long long ANumArgs);
 			public:
 				template <typename... Args> static unsigned long long Call(const CallNumber ACallNumber, Args&&... AArgs) {
 					VerifyArgumentCount(ACallNumber, sizeof...(AArgs));
